Migrated callback from devgroup to kdn
